+++ src/main/java/inu/swcontest/gm/service/ImageService.java
@@ -1,7 +1,6 @@
-package inu.swcontest.gm.Service;
+package inu.swcontest.gm.service;
 
 import inu.swcontest.gm.model.UploadImageRequest;
-import org.springframework.web.multipart.MultipartFile;
 
 public interface ImageService {
     void uploadImage(UploadImageRequest request);
