+++ src/main/java/inu/swcontest/gm/service/serviceImpl/ImageServiceImpl.java
@@ -1,14 +1,13 @@
-package inu.swcontest.gm.Service.impl;
+package inu.swcontest.gm.service.serviceImpl;
 
 import com.google.cloud.storage.BlobInfo;
 import com.google.cloud.storage.Storage;
 import com.google.cloud.storage.StorageOptions;
-import inu.swcontest.gm.Service.ImageService;
+import inu.swcontest.gm.service.ImageService;
 import inu.swcontest.gm.model.UploadImageRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.UUID;
@@ -34,7 +33,6 @@ public class ImageServiceImpl implements ImageService {
         String contentType = request.getImage().getContentType();
 
         // 이미지 정보 설정
-
         try {
             BlobInfo blobInfo = BlobInfo.newBuilder(bucketName, name).setContentType(contentType).build();
